requirements.txt
wsgidav>=4.3.0
requests>=2.31.0
parse-torrent-title>=2.8.0
python-dotenv>=1.0.0
cachetools>=5.3.0
cheroot>=11.0.0
